chore: remove obsolete vote files
