+++ pytest.ini
@@ -3,3 +3,4 @@ DJANGO_SETTINGS_MODULE = config.test_settings
 testpaths =
     tests
     core/tests
+    votes/tests
