+++ templates/votes/partials/vote_widget.html
