+++ votes/admin.py
@@ -1,0 +1,8 @@
+from django.contrib import admin
+
+from .models import Vote
+
+
+@admin.register(Vote)
+class VoteAdmin(admin.ModelAdmin):
+    list_display = ("id", "user", "target_type", "target_id", "value")
